post message text includes creation date and artist, improve image description
So far the image description only contained the link to the source url. This was not very accessibility friendly as people with visual impairment would not get a proper description of the posted image. Meanwhile, for people who want to get to the source page couldn't do so, as they would need to copy the URL from the former alt text.
With this change, the image author, creation time and source url are provided in the post text and the image description given by the API is posted as alt text.